+++ demos/widgets/lv_demo_widgets_shop.h
@@ -1,0 +1,39 @@
+/**
+ * @file lv_demo_widgets_shop.h
+ *
+ */
+
+#ifndef LV_DEMO_WIDGETS_SHOP_H
+#define LV_DEMO_WIDGETS_SHOP_H
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+/*********************
+ *      INCLUDES
+ *********************/
+#include "../lv_demos.h"
+
+/*********************
+ *      DEFINES
+ *********************/
+
+/**********************
+ *      TYPEDEFS
+ **********************/
+
+/**********************
+ * GLOBAL PROTOTYPES
+ **********************/
+void lv_demo_widgets_shop_create(lv_obj_t * parent);
+
+/**********************
+ *      MACROS
+ **********************/
+
+#ifdef __cplusplus
+} /*extern "C"*/
+#endif
+
+#endif /*LV_DEMO_WIDGETS_SHOP_H*/
